+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Go Buildpack Changelog
 
 ## Unreleased
+* Drop cedar-14 from test matrix
+* Remove skipping of tests on cedar-14
+* Update Makefile's default IMAGE to heroku/heroku:20-build
 
 ## v151 (2021-02-01)
 * Add go1.16rc1, use for go1.16